requirements.txt
#requirements.txt
psycopg2-binary==2.9.10
minio==7.2.16
requests==2.32.5
sqlalchemy

dbt-core==1.10.11
dbt-spark[PyHive]
dbt-trino==1.9.3

scrapy==2.13.3

pyspark==3.5.6
delta-spark==3.1.0

trino==0.336.0

apache-airflow-providers-trino==6.3.3
apache-airflow-providers-apache-spark==5.3.2